+++ firmware/config/boards/subaru_eg33/config/controllers/algo/engine_configuration_generated_structures.h
@@ -1,4 +1,4 @@
-// this section was generated automatically by rusEFI tool ConfigDefinition.jar based on config/boards/subaru_eg33/config/gen_config.sh integration/rusefi_config.txt Fri Aug 27 09:53:14 UTC 2021
+// this section was generated automatically by rusEFI tool ConfigDefinition.jar based on config/boards/subaru_eg33/config/gen_config.sh integration/rusefi_config.txt Sun Sep 19 01:14:49 UTC 2021
 // by class com.rusefi.output.CHeaderConsumer
 // begin
 #pragma once
@@ -189,13 +189,13 @@ struct gppwm_channel {
 	 */
 	uint16_t pwmFrequency;
 	/**
-	 * In on-off mode, turn the output on when the table value is above this duty.
+	 * Hysteresis: in on-off mode, turn the output on when the table value is above this duty.
 	%
 	 * offset 4
 	 */
 	uint8_t onAboveDuty;
 	/**
-	 * In on-off mode, turn the output off when the table value is below this duty.
+	 * Hysteresis: in on-off mode, turn the output off when the table value is below this duty.
 	%
 	 * offset 5
 	 */
@@ -797,7 +797,7 @@ struct engine_configuration_s {
 	bool issue_294_29 : 1;
 	/**
 	offset 76 bit 29 */
-	bool issue_294_30 : 1;
+	bool artificialTestMisfire : 1;
 	/**
 	offset 76 bit 30 */
 	bool issue_294_31 : 1;
@@ -1664,9 +1664,11 @@ struct engine_configuration_s {
 	 */
 	uint8_t mc33_hvolt;
 	/**
+	 * Minimum MAP before closed loop boost is enabled. Use to prevent misbehavior upon entering boost.
+	kPa
 	 * offset 761
 	 */
-	uint8_t unused761;
+	uint8_t minimumBoostClosedLoopMap;
 	/**
 	 * Optional Radiator Fan used with A/C
 	 * offset 762
@@ -1681,22 +1683,28 @@ struct engine_configuration_s {
 	 */
 	pin_output_mode_e gpioPinModes[FSIO_COMMAND_COUNT];
 	/**
+	volts
 	 * offset 770
 	 */
-	uint8_t unusedpinModesWhereHere[2];
+	uint8_t dwellVoltageCorrVoltBins[DWELL_CURVE_SIZE];
 	/**
-	 * offset 772
+	 * offset 778
 	 */
-	adc_channel_e luaAnalogInputs[LUA_ANALOG_INPUT_COUNT];
+	uint8_t unusedpinModesWhereHere[2];
 	/**
 	 * todo: more comments
 	 * offset 780
 	 */
 	output_pin_e fsioOutputPins[FSIO_COMMAND_COUNT];
 	/**
+	multiplier
 	 * offset 786
 	 */
-	uint8_t unusedOutputWhereHere[10];
+	uint8_t dwellVoltageCorrValues[DWELL_CURVE_SIZE];
+	/**
+	 * offset 794
+	 */
+	uint8_t unusedOutputWhereHere[2];
 	/**
 	 * offset 796
 	 */
@@ -2023,10 +2031,10 @@ struct engine_configuration_s {
 	bool unusedBit_251_29 : 1;
 	/**
 	offset 976 bit 30 */
-	bool unusedBit_297_30 : 1;
+	bool unusedBit_298_30 : 1;
 	/**
 	offset 976 bit 31 */
-	bool unusedBit_297_31 : 1;
+	bool unusedBit_298_31 : 1;
 	/**
 	 * offset 980
 	 */
@@ -2194,7 +2202,7 @@ struct engine_configuration_s {
 	/**
 	 * offset 1244
 	 */
-	vr_threshold_s vrThreshold[2];
+	vr_threshold_s vrThreshold[VR_THRESHOLD_COUNT];
 	/**
 	units
 	 * offset 1276
@@ -2986,76 +2994,76 @@ struct engine_configuration_s {
 	bool unused1130 : 1;
 	/**
 	offset 2116 bit 8 */
-	bool unusedBit_507_8 : 1;
+	bool unusedBit_508_8 : 1;
 	/**
 	offset 2116 bit 9 */
-	bool unusedBit_507_9 : 1;
+	bool unusedBit_508_9 : 1;
 	/**
 	offset 2116 bit 10 */
-	bool unusedBit_507_10 : 1;
+	bool unusedBit_508_10 : 1;
 	/**
 	offset 2116 bit 11 */
-	bool unusedBit_507_11 : 1;
+	bool unusedBit_508_11 : 1;
 	/**
 	offset 2116 bit 12 */
-	bool unusedBit_507_12 : 1;
+	bool unusedBit_508_12 : 1;
 	/**
 	offset 2116 bit 13 */
-	bool unusedBit_507_13 : 1;
+	bool unusedBit_508_13 : 1;
 	/**
 	offset 2116 bit 14 */
-	bool unusedBit_507_14 : 1;
+	bool unusedBit_508_14 : 1;
 	/**
 	offset 2116 bit 15 */
-	bool unusedBit_507_15 : 1;
+	bool unusedBit_508_15 : 1;
 	/**
 	offset 2116 bit 16 */
-	bool unusedBit_507_16 : 1;
+	bool unusedBit_508_16 : 1;
 	/**
 	offset 2116 bit 17 */
-	bool unusedBit_507_17 : 1;
+	bool unusedBit_508_17 : 1;
 	/**
 	offset 2116 bit 18 */
-	bool unusedBit_507_18 : 1;
+	bool unusedBit_508_18 : 1;
 	/**
 	offset 2116 bit 19 */
-	bool unusedBit_507_19 : 1;
+	bool unusedBit_508_19 : 1;
 	/**
 	offset 2116 bit 20 */
-	bool unusedBit_507_20 : 1;
+	bool unusedBit_508_20 : 1;
 	/**
 	offset 2116 bit 21 */
-	bool unusedBit_507_21 : 1;
+	bool unusedBit_508_21 : 1;
 	/**
 	offset 2116 bit 22 */
-	bool unusedBit_507_22 : 1;
+	bool unusedBit_508_22 : 1;
 	/**
 	offset 2116 bit 23 */
-	bool unusedBit_507_23 : 1;
+	bool unusedBit_508_23 : 1;
 	/**
 	offset 2116 bit 24 */
-	bool unusedBit_507_24 : 1;
+	bool unusedBit_508_24 : 1;
 	/**
 	offset 2116 bit 25 */
-	bool unusedBit_507_25 : 1;
+	bool unusedBit_508_25 : 1;
 	/**
 	offset 2116 bit 26 */
-	bool unusedBit_507_26 : 1;
+	bool unusedBit_508_26 : 1;
 	/**
 	offset 2116 bit 27 */
-	bool unusedBit_507_27 : 1;
+	bool unusedBit_508_27 : 1;
 	/**
 	offset 2116 bit 28 */
-	bool unusedBit_507_28 : 1;
+	bool unusedBit_508_28 : 1;
 	/**
 	offset 2116 bit 29 */
-	bool unusedBit_507_29 : 1;
+	bool unusedBit_508_29 : 1;
 	/**
 	offset 2116 bit 30 */
-	bool unusedBit_507_30 : 1;
+	bool unusedBit_508_30 : 1;
 	/**
 	offset 2116 bit 31 */
-	bool unusedBit_507_31 : 1;
+	bool unusedBit_508_31 : 1;
 	/**
 	 * set can_mode X
 	 * offset 2120
@@ -3119,7 +3127,7 @@ struct engine_configuration_s {
 	 * todo: rename to fsioAnalogInputs
 	 * offset 2200
 	 */
-	adc_channel_e fsioAdc[FSIO_ANALOG_INPUT_COUNT];
+	adc_channel_e fsioAdc[AUX_ANALOG_INPUT_COUNT];
 	/**
 	 * Fixed timing, useful for TDC testing
 	deg
@@ -4335,4 +4343,4 @@ struct persistent_config_s {
 };
 
 // end
-// this section was generated automatically by rusEFI tool ConfigDefinition.jar based on config/boards/subaru_eg33/config/gen_config.sh integration/rusefi_config.txt Fri Aug 27 09:53:14 UTC 2021
+// this section was generated automatically by rusEFI tool ConfigDefinition.jar based on config/boards/subaru_eg33/config/gen_config.sh integration/rusefi_config.txt Sun Sep 19 01:14:49 UTC 2021