+++ firmware/controllers/actuators/idle_thread.cpp
@@ -34,7 +34,6 @@
 #include "idle_hardware.h"
 
 #include "periodic_task.h"
-#include "vehicle_speed.h"
 #include "dc_motors.h"
 
 #if EFI_TUNER_STUDIO