+++ firmware/controllers/serial/serial_rx.cpp
@@ -14,8 +14,6 @@
 #include "serial_hw.h"
 #include "serial_sensor.h"
 
-#include "vehicle_speed.h"
-
 uint8_t ser_buffer[SERBUFFLEN] = {};
 size_t innovate_msg_len = 1;
 innovate_serial_id_state_t innovate_serial_id_state = UNKNOWN;