+++ java_console/ui/src/main/java/com/rusefi/maintenance/DfuFlasher.java
@@ -18,6 +18,8 @@ import java.awt.event.ActionListener;
 import java.io.File;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import static com.rusefi.StartupFrame.appendBundleName;
+
 /**
  * @see FirmwareFlasher
  */
@@ -69,14 +71,14 @@ public class DfuFlasher {
             }
         }
         StatusWindow wnd = new StatusWindow();
-        wnd.showFrame("DFU status " + Launcher.CONSOLE_VERSION);
+        wnd.showFrame(appendBundleName("DFU status " + Launcher.CONSOLE_VERSION));
         wnd.appendMsg(messages.toString());
         ExecHelper.submitAction(() -> executeDFU(wnd), DfuFlasher.class + " thread");
     }
 
     public static void runDfuProgramming() {
         StatusWindow wnd = new StatusWindow();
-        wnd.showFrame("DFU status " + Launcher.CONSOLE_VERSION);
+        wnd.showFrame(appendBundleName("DFU status " + Launcher.CONSOLE_VERSION));
         ExecHelper.submitAction(() -> executeDFU(wnd), DfuFlasher.class + " thread");
     }
 
@@ -105,7 +107,11 @@ public class DfuFlasher {
                     }
                     wnd.appendMsg(s);
                 }, stdout);
-        if (stdout.toString().contains("Verify successful") || stdout.toString().contains("Upgrade successful")) {
+        if (stdout.toString().contains("Matching not good")) {
+            // looks like sometimes we are not catching the last line of the response? 'Upgrade' happens before 'Verify'
+            wnd.appendMsg("VERIFICATION ERROR maybe nDBANK issue?");
+            wnd.appendMsg("https://github.com/rusefi/rusefi/wiki/HOWTO-nDBANK");
+        } else if (stdout.toString().contains("Verify successful") || stdout.toString().contains("Upgrade successful")) {
             // looks like sometimes we are not catching the last line of the response? 'Upgrade' happens before 'Verify'
             wnd.appendMsg("SUCCESS!");
         } else {