+++ java_console/ui/src/main/java/com/rusefi/ui/lua/LuaScriptPanel.java
@@ -1,7 +1,6 @@
 package com.rusefi.ui.lua;
 
 import com.opensr5.ConfigurationImage;
-import com.rusefi.FixedCommandControl;
 import com.rusefi.binaryprotocol.BinaryProtocol;
 import com.rusefi.config.generated.Fields;
 import com.rusefi.ui.MessagesPanel;
@@ -10,24 +9,23 @@ import com.rusefi.ui.storage.Node;
 import com.rusefi.ui.widgets.AnyCommand;
 
 import javax.swing.*;
-import javax.swing.border.Border;
 import java.awt.*;
-import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.nio.ByteBuffer;
-import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
+
+import static com.rusefi.ui.util.UiUtils.trueLayout;
 
 public class LuaScriptPanel {
     private final UIContext context;
     private final JPanel mainPanel = new JPanel(new BorderLayout());
     private final AnyCommand command;
-    private final JTextArea scriptText;
+    private final JTextArea scriptText = new JTextArea();
+    private boolean isFirstRender = true;
 
     public LuaScriptPanel(UIContext context, Node config) {
         this.context = context;
-        this.command = AnyCommand.createField(context, config, true, true);
+        command = AnyCommand.createField(context, config, true, true);
 
         // Upper panel: command entry, etc
         JPanel upperPanel = new JPanel(new FlowLayout(FlowLayout.LEFT, 5, 0));
@@ -36,20 +34,19 @@ public class LuaScriptPanel {
         JButton writeButton = new JButton("Write to ECU");
         JButton resetButton = new JButton("Reset/Reload Lua");
 
-        readButton.addActionListener(e -> read());
+        readButton.addActionListener(e -> readFromECU());
         writeButton.addActionListener(e -> write());
         resetButton.addActionListener(e -> resetLua());
 
         upperPanel.add(readButton);
         upperPanel.add(writeButton);
         upperPanel.add(resetButton);
-        upperPanel.add(this.command.getContent());
+        upperPanel.add(command.getContent());
 
         // Center panel - script editor and log
         JPanel scriptPanel = new JPanel(new BorderLayout());
-        this.scriptText = new JTextArea();
-        this.scriptText.setTabSize(2);
-        scriptPanel.add(this.scriptText, BorderLayout.CENTER);
+        scriptText.setTabSize(2);
+        scriptPanel.add(scriptText, BorderLayout.CENTER);
 
         //centerPanel.add(, BorderLayout.WEST);
         JPanel messagesPanel = new JPanel(new BorderLayout());
@@ -57,10 +54,27 @@ public class LuaScriptPanel {
         messagesPanel.add(BorderLayout.NORTH, mp.getButtonPanel());
         messagesPanel.add(BorderLayout.CENTER, mp.getMessagesScroll());
 
-        JSplitPane centerPanel = new JSplitPane(JSplitPane.HORIZONTAL_SPLIT, scriptPanel, messagesPanel);
+        JSplitPane centerPanel = new JSplitPane(JSplitPane.HORIZONTAL_SPLIT, scriptPanel, messagesPanel) {
+            @Override
+            public void paint(Graphics g) {
+                super.paint(g);
+                if (isFirstRender) {
+                    readFromECU();
+                    isFirstRender = true;
+                }
+            }
+        };
 
-        this.mainPanel.add(upperPanel, BorderLayout.NORTH);
-        this.mainPanel.add(centerPanel, BorderLayout.CENTER);
+        mainPanel.add(upperPanel, BorderLayout.NORTH);
+        mainPanel.add(centerPanel, BorderLayout.CENTER);
+
+        trueLayout(mainPanel);
+        SwingUtilities.invokeLater(new Runnable() {
+            @Override
+            public void run() {
+                centerPanel.setDividerLocation(centerPanel.getSize().width / 2);
+            }
+        });
     }
 
     public JPanel getPanel() {
@@ -68,33 +82,39 @@ public class LuaScriptPanel {
     }
 
     public ActionListener getTabSelectedListener() {
-        return new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                if (command != null)
-                    command.requestFocus();
-            }
+        return e -> {
+            if (command != null)
+                command.requestFocus();
         };
     }
 
-    void read() {
-        BinaryProtocol bp = this.context.getLinkManager().getCurrentStreamState();
+    void readFromECU() {
+        BinaryProtocol bp = context.getLinkManager().getCurrentStreamState();
 
         if (bp == null) {
-            // TODO: Handle missing ECU
+            scriptText.setText("No ECU located");
             return;
         }
 
         ConfigurationImage image = bp.getControllerConfiguration();
+        if (image == null) {
+            scriptText.setText("No configuration image");
+            return;
+        }
         ByteBuffer luaScriptBuffer = image.getByteBuffer(Fields.luaScript_offset, Fields.LUA_SCRIPT_SIZE);
 
-        byte scriptArr[] = new byte[Fields.LUA_SCRIPT_SIZE];
+        byte[] scriptArr = new byte[Fields.LUA_SCRIPT_SIZE];
         luaScriptBuffer.get(scriptArr);
 
+        int i = findNullTerminator(scriptArr);
+        scriptText.setText(new String(scriptArr, 0, i, StandardCharsets.US_ASCII));
+    }
+
+    @SuppressWarnings("StatementWithEmptyBody")
+    private static int findNullTerminator(byte[] scriptArr) {
         int i;
-        // Find the null terminator
         for (i = 0; i < scriptArr.length && scriptArr[i] != 0; i++) ;
-        scriptText.setText(new String(scriptArr, 0, i, Charset.forName("ASCII")));
+        return i;
     }
 
     void write() {
@@ -102,8 +122,8 @@ public class LuaScriptPanel {
 
         String script = scriptText.getText();
 
-        byte paddedScript[] = new byte[Fields.LUA_SCRIPT_SIZE];
-        byte scriptBytes[] = script.getBytes(StandardCharsets.US_ASCII);
+        byte[] paddedScript = new byte[Fields.LUA_SCRIPT_SIZE];
+        byte[] scriptBytes = script.getBytes(StandardCharsets.US_ASCII);
         System.arraycopy(scriptBytes, 0, paddedScript, 0, scriptBytes.length);
 
         int idx = 0;
@@ -111,7 +131,7 @@ public class LuaScriptPanel {
 
         do {
             remaining = paddedScript.length - idx;
-            int thisWrite = remaining > Fields.BLOCKING_FACTOR ? Fields.BLOCKING_FACTOR : remaining;
+            int thisWrite = Math.min(remaining, Fields.BLOCKING_FACTOR);
 
             bp.writeData(paddedScript, idx, Fields.luaScript_offset + idx, thisWrite);