+++ unit_tests/tests/test_launch.cpp
@@ -1,7 +1,6 @@
 #include "pch.h"
 
 #include "launch_control.h"
-#include "vehicle_speed.h"
 
 TEST(LaunchControl, TpsCondition) {
 	WITH_ENGINE_TEST_HELPER(TEST_ENGINE);